requirements.txt
streamlit
rank-bm25
transformers
torch
nltk
PyPDF2
groq
faiss-cpu
gensim
numpy